% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_precip_by_month.R
\name{mean_precip_by_month}
\alias{mean_precip_by_month}
\title{Mean precipitation for each water year}
\usage{
mean_precip_by_month(precip_array, plot = TRUE)
}
\arguments{
\item{precip_array}{array that contains the following information}
}
\value{
what your function returns (outputs)
}
\description{
This function calculates the mean precipation by month for all years and produces a graph to visualize the data.
}
\references{
Source: Monthly Observed Precipitation - NWS Cooperative Observers. The following are data from NWS cooperative observers for the current water year and historically back to Water Year 2002. The water year starts on Oct 1 and ends on Sept 30. Precipitation data is provided for each month of the current water year, total precipitation for the water year, the percent of normal for the water year to date, and the percent of the entire water year received to date. https://www.cnrfc.noaa.gov/rainfall_data.php#monthly
}
\author{
Jamie Miller and Anna Calle
}
