% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_precip_by_season.R
\name{mean_precip_by_season}
\alias{mean_precip_by_season}
\title{Mean Precipitation per Season}
\usage{
mean_precip_by_season(precip_data, year)
}
\arguments{
\item{precip_data}{Data frame that contains the following variables: water_year, Location, ID, month, and precip}

\item{year}{Year selected}
}
\value{
Table of each season and its mean precipitation for the selected year
}
\description{
This function calculates the mean precipation by season for a given year
}
\references{
citations or urls
}
\author{
Jamie Miller and Anna Calle
}
