% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_array.R
\docType{data}
\name{precip_array}
\alias{precip_array}
\title{Monthly Precipitation Array}
\format{An object of class \code{array} of dimension 18 x 3 x 12.}
\usage{
precip_array
}
\value{
Returns precip_array.
}
\description{
This function creates an array from the monthly_precip$precip data frame, water_year, Location, and month.
}
\references{
Source: Monthly Observed Precipitation - NWS Cooperative Observers. The following are data from NWS cooperative observers for the current water year and historically back to Water Year 2002. The water year starts on Oct 1 and ends on Sept 30. Precipitation data is provided for each month of the current water year, total precipitation for the water year, the percent of normal for the water year to date, and the percent of the entire water year received to date. \url{https://www.cnrfc.noaa.gov/rainfall_data.php#monthly}
}
\author{
Jamie Miller and Anna Calle
}
\keyword{datasets}
