% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_precip_by_year.R
\name{total_precip_by_year}
\alias{total_precip_by_year}
\title{Total precipitation for a each water year and location}
\usage{
total_precip_by_year(precip_data, plot = FALSE)
}
\arguments{
\item{precip_data}{The function uses data from the precipitaion data frame "monthly_precip".}
}
\value{
Returns either a graph or table that summarizes the total annual precipitation for each water year and location.
}
\description{
This function calculates the total precipation for each year at a given location and produces a graph to visualize the data.
}
\references{
Source: Monthly Observed Precipitation - NWS Cooperative Observers. The following are data from NWS cooperative observers for the current water year and historically back to Water Year 2002. The water year starts on Oct 1 and ends on Sept 30. Precipitation data is provided for each month of the current water year, total precipitation for the water year, the percent of normal for the water year to date, and the percent of the entire water year received to date. \url{https://www.cnrfc.noaa.gov/rainfall_data.php#monthly}
}
\author{
Jamie Miller and Anna Calle
}
