% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_water_tax.R
\name{calc_water_tax}
\alias{calc_water_tax}
\title{Calculate Water Tax}
\usage{
calc_water_tax(precip_data, household_consumption, location, month_tax,
  year)
}
\arguments{
\item{precip_data}{Data frame that contains the following variables: water_year, Location, ID, month, and precip}

\item{household_consumption}{Average household water consumption in gallons. Average in California is 86 gallons}

\item{location}{One of three locations: "Paso Robles", "San Luis Obispo", or "Santa Barbara"}

\item{month_tax}{The three-letter month code}

\item{year}{The water year 2002 through 2019}
}
\value{
Tax price in dollars
}
\description{
This function calculates the monthly tax on consumers for safe drinking water. The tax depends on household water consumption as well as availability of clean water sources. During low precipitation months the city has to turn to more expensive sources of water like desalination or imported water, raising the price of the tax.
}
\references{
citations or urls
}
\author{
Jamie Miller and Anna Calle
}
