% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_flooding.R
\name{predict_flooding}
\alias{predict_flooding}
\title{Predict Flooding Risk}
\usage{
predict_flooding(precip_data, year)
}
\arguments{
\item{precip_data}{data frame that contains the following information}
}
\value{
what your function returns (outputs)
}
\description{
This function predict for a given year
}
\references{
citations or urls
}
\author{
Jamie Miller and Anna Calle
}
