% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_flooding.R
\name{predict_flooding}
\alias{predict_flooding}
\title{Predict Flooding Risk}
\usage{
predict_flooding(precip_data, year)
}
\arguments{
\item{precip_data}{Data frame that contains the following variables: water_year, Location, ID, month, and precip.}

\item{year}{Year selected}
}
\value{
Location, month, mean historic precipitation, precipitation and difference between mean historic precipitation and precipitation of place with highest risk of flooding
}
\description{
This function predicts the location and month that had the highest risk of flooding in a given year. This is determined by calculating the location and month that have the greatest positive difference from the historic precipitation average.
}
\references{
citations or urls
}
\author{
Jamie Miller and Anna Calle
}
